first working game ECS

+++ PositionComponent.hpp
@@ -13,7 +13,7 @@ public:
     ~PositionComponent();
     //accesors
     const sf::Vector2f& getPosition();
-    const float& getAngle();
+    const double& getAngle();
     //functions
     void move(const double dx, const double dy, const double dAngle);
     void rotate(const double dAngle);
